refactor: remove BlockChainState
