+++ Libplanet/Blockchain/BlockChain.cs
@@ -17,6 +17,7 @@ using Libplanet.Blockchain.Renderers;
 using Libplanet.Common;
 using Libplanet.Crypto;
 using Libplanet.Store;
+using Libplanet.Store.Trie;
 using Libplanet.Types.Assets;
 using Libplanet.Types.Blocks;
 using Libplanet.Types.Consensus;
@@ -34,7 +35,7 @@ namespace Libplanet.Blockchain
     /// <para>
     /// In order to watch its state changes, implement <see cref="IRenderer"/> interface
     /// and pass it to the
-    /// <see cref="BlockChain(IBlockPolicy, IStagePolicy, IStore, IStateStore, Block, IBlockChainStates, IActionEvaluator, IEnumerable{IRenderer})"/>
+    /// <see cref="BlockChain(IBlockPolicy, IStagePolicy, IStore, IStateStore, Block, IActionEvaluator, IEnumerable{IRenderer})"/>
     /// constructor.
     /// </para>
     /// </summary>
@@ -51,7 +52,6 @@ namespace Libplanet.Blockchain
         internal readonly ReaderWriterLockSlim _rwlock;
         private readonly object _txLock;
         private readonly ILogger _logger;
-        private readonly IBlockChainStates _blockChainStates;
 
         /// <summary>
         /// All <see cref="Block"/>s in the <see cref="BlockChain"/>
@@ -81,8 +81,6 @@ namespace Libplanet.Blockchain
         /// it checks if the existing genesis block and this argument is the same.
         /// If the <paramref name="store"/> has no genesis block yet this argument will
         /// be used for that.</param>
-        /// <param name="blockChainStates">The <see cref="IBlockChainStates"/> implementation for
-        /// state lookup.</param>
         /// <param name="actionEvaluator">The <see cref="ActionEvaluator" /> implementation to
         /// calculate next states when append new blocks.</param>
         /// <param name="renderers">Listens state changes on the created chain.  Listens nothing
@@ -104,7 +102,6 @@ namespace Libplanet.Blockchain
             IStore store,
             IStateStore stateStore,
             Block genesisBlock,
-            IBlockChainStates blockChainStates,
             IActionEvaluator actionEvaluator,
             IEnumerable<IRenderer> renderers = null)
 #pragma warning disable SA1118  // The parameter spans multiple lines
@@ -118,7 +115,6 @@ namespace Libplanet.Blockchain
                         $"Given {nameof(store)} does not have canonical chain id set.",
                         nameof(store)),
                 genesisBlock,
-                blockChainStates,
                 actionEvaluator,
                 renderers)
         {
@@ -131,7 +127,6 @@ namespace Libplanet.Blockchain
             IStateStore stateStore,
             Guid id,
             Block genesisBlock,
-            IBlockChainStates blockChainStates,
             IActionEvaluator actionEvaluator,
             IEnumerable<IRenderer> renderers)
         {
@@ -154,8 +149,6 @@ namespace Libplanet.Blockchain
             StagePolicy = stagePolicy;
             Store = store;
             StateStore = stateStore;
-
-            _blockChainStates = blockChainStates;
 
             _blocks = new BlockSet(store);
             Renderers = renderers is IEnumerable<IRenderer> r
@@ -206,7 +199,7 @@ namespace Libplanet.Blockchain
         /// <remarks>
         /// Since this value is immutable, renderers cannot be registered after once a <see
         /// cref="BlockChain"/> object is instantiated; use <c>renderers</c> option of
-        /// <see cref="BlockChain(IBlockPolicy, IStagePolicy, IStore, IStateStore, Block, IBlockChainStates, IActionEvaluator, IEnumerable{IRenderer})"/>
+        /// <see cref="BlockChain(IBlockPolicy, IStagePolicy, IStore, IStateStore, Block, IActionEvaluator, IEnumerable{IRenderer})"/>
         /// constructor instead.
         /// </remarks>
  #pragma warning restore MEN002
@@ -353,8 +346,7 @@ namespace Libplanet.Blockchain
             IStateStore stateStore,
             Block genesisBlock,
             IActionEvaluator actionEvaluator,
-            IEnumerable<IRenderer> renderers = null,
-            IBlockChainStates blockChainStates = null)
+            IEnumerable<IRenderer> renderers = null)
 #pragma warning restore SA1611  // The documentation for parameters are missing.
         {
             if (store is null)
@@ -414,8 +406,6 @@ namespace Libplanet.Blockchain
 
             store.SetCanonicalChainId(id);
 
-            blockChainStates ??= new BlockChainStates(store, stateStore);
-
             return new BlockChain(
                 policy,
                 stagePolicy,
@@ -423,7 +413,6 @@ namespace Libplanet.Blockchain
                 stateStore,
                 id,
                 genesisBlock,
-                blockChainStates,
                 actionEvaluator,
                 renderers);
         }
@@ -499,7 +488,7 @@ namespace Libplanet.Blockchain
 
         /// <inheritdoc cref="IBlockChainStates.GetState"/>
         public IValue GetState(Address address, BlockHash? offset) =>
-            _blockChainStates.GetState(address, offset);
+            GetAccountState(offset).GetState(address);
 
         /// <summary>
         /// Gets multiple states associated to the specified <paramref name="addresses"/>.
@@ -516,7 +505,7 @@ namespace Libplanet.Blockchain
         public IReadOnlyList<IValue> GetStates(
             IReadOnlyList<Address> addresses,
             BlockHash? offset) =>
-            _blockChainStates.GetStates(addresses, offset);
+            GetAccountState(offset).GetStates(addresses);
 
         /// <summary>
         /// Queries <paramref name="address"/>'s current balance of the <paramref name="currency"/>
@@ -536,7 +525,7 @@ namespace Libplanet.Blockchain
             Address address,
             Currency currency,
             BlockHash? offset) =>
-            _blockChainStates.GetBalance(address, currency, offset);
+            GetAccountState(offset).GetBalance(address, currency);
 
         /// <summary>
         /// Gets the current total supply of a <paramref name="currency"/> in the
@@ -550,21 +539,21 @@ namespace Libplanet.Blockchain
 
         /// <inheritdoc cref="IBlockChainStates.GetTotalSupply"/>
         public FungibleAssetValue GetTotalSupply(Currency currency, BlockHash? offset) =>
-            _blockChainStates.GetTotalSupply(currency, offset);
+            GetAccountState(offset).GetTotalSupply(currency);
 
         public ValidatorSet GetValidatorSet() => GetValidatorSet(Tip.Hash);
 
         /// <inheritdoc cref="IBlockChainStates.GetValidatorSet" />
         public ValidatorSet GetValidatorSet(BlockHash? offset) =>
-            _blockChainStates.GetValidatorSet(offset);
+            GetAccountState(offset).GetValidatorSet();
 
         /// <inheritdoc cref="IBlockChainStates.GetAccountState(BlockHash?)" />
         public IAccountState GetAccountState(BlockHash? offset) =>
-            _blockChainStates.GetAccountState(offset);
+            new AccountState(GetTrie(offset));
 
         /// <inheritdoc cref="IBlockChainStates.GetAccountState(HashDigest{SHA256}?)" />
         public IAccountState GetAccountState(HashDigest<SHA256>? hash) =>
-            _blockChainStates.GetAccountState(hash);
+            new AccountState(GetTrie(hash));
 
         /// <summary>
         /// Queries the recorded <see cref="TxExecution"/> for a successful or failed
@@ -829,7 +818,6 @@ namespace Libplanet.Blockchain
                     StateStore,
                     forkedId,
                     Genesis,
-                    _blockChainStates,
                     ActionEvaluator,
                     renderers);
                 Store.ForkBlockIndexes(Id, forkedId, point);
@@ -1285,6 +1273,58 @@ namespace Libplanet.Blockchain
                     Store.DeleteBlockCommit(hash);
                 }
             }
+        }
+
+        /// <summary>
+        /// Returns the state root associated with <see cref="BlockHash"/>
+        /// <paramref name="offset"/>.
+        /// </summary>
+        /// <param name="offset">The <see cref="BlockHash"/> to look up in
+        /// the internally held <see cref="IStore"/>.</param>
+        /// <returns>An <see cref="ITrie"/> representing the state root associated with
+        /// <paramref name="offset"/>.</returns>
+        /// <exception cref="ArgumentException">Thrown for one of the following reasons.
+        /// <list type="bullet">
+        ///     <item><description>
+        ///         If <paramref name="offset"/> is not <see langword="null"/> and
+        ///         <paramref name="offset"/> cannot be found in <see cref="IStore"/>.
+        ///     </description></item>
+        ///     <item><description>
+        ///         If <paramref name="offset"/> is not <see langword="null"/> and
+        ///         the state root hash associated with <paramref name="offset"/>
+        ///         cannot be found in <see cref="IStateStore"/>.
+        ///     </description></item>
+        /// </list>
+        /// </exception>
+        /// <remarks>
+        /// An <see cref="ITrie"/> returned by this method is read-only.
+        /// </remarks>
+        private ITrie GetTrie(BlockHash? offset)
+        {
+            if (!(offset is { } hash))
+            {
+                return StateStore.GetStateRoot(null);
+            }
+            else if (Store.GetStateRootHash(hash) is { } stateRootHash)
+            {
+                return GetTrie(stateRootHash);
+            }
+            else
+            {
+                throw new ArgumentException(
+                    $"Could not find block hash {hash} in {nameof(IStore)}.",
+                    nameof(offset));
+            }
+        }
+
+        private ITrie GetTrie(HashDigest<SHA256>? hash)
+        {
+            ITrie trie = StateStore.GetStateRoot(hash);
+            return trie.Recorded
+                ? trie
+                : throw new ArgumentException(
+                    $"Could not find state root {hash} in {nameof(IStateStore)}.",
+                    nameof(hash));
         }
     }
 }
